refactor: rename request module to api for clarity
Rename the `request` module to `api` in the file paths to more accurately reflect its functionality. Update all references from `request` to `api` accordingly. Adjust related documentation and test cases to align with the new module name.
