+++ README.md
@@ -6,8 +6,8 @@
 
 ## Features
 
-1. 支持本地变量命名风格转换
-2. 支持 SiliconCloud API 变量命名风格转换 (实验)
+1. 支持基于 change-case-all 变量命名风格转换
+2. 支持基于 SiliconCloud 变量命名风格转换 (实验, 根据目前的提示词输出不太稳定)
 
 ## Usage
